+++ app/controllers/users_controller.rb
@@ -2,6 +2,6 @@ class UsersController < ApplicationController
   before_action :require_admin
 
   def index
-    @users = User.alphabetical.includes(owned_jobs: :printer)
+    @users = User.in_display_name_order(User.includes(owned_jobs: :printer))
   end
 end